refact: ChooseMyType ViewModel / Repository 파일 분할
