+++ Tlog/app/src/main/java/com/tlog/viewmodel/beginning/ChooseMyTypeViewModel.kt
@@ -2,19 +2,16 @@ package com.tlog.viewmodel.beginning
 
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
-import com.tlog.api.LoginApi
 import com.tlog.api.retrofit.TokenProvider
-import com.tlog.data.api.BaseResponse
 import com.tlog.data.api.FcmTokenBody
-import com.tlog.data.api.FirebaseTokenData
 import com.tlog.data.api.RegisterRequest
 import com.tlog.data.api.UserProfileDto
 import com.tlog.data.local.UserPreferences
+import com.tlog.data.repository.ChooseMyTypeRepository
 import com.tlog.ui.navigation.Screen
 import com.tlog.viewmodel.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import jakarta.inject.Inject
-import retrofit2.Response
 
 @HiltViewModel
 class ChooseMyTypeViewModel @Inject constructor(
@@ -75,18 +72,5 @@ class ChooseMyTypeViewModel @Inject constructor(
                 }
             }
         )
-    }
-}
-
-
-class ChooseMyTypeRepository @Inject constructor(
-    private val loginRetrofitInstance: LoginApi
-) {
-    suspend fun ssoRegister(request: RegisterRequest): Response<BaseResponse<FirebaseTokenData>> {
-        return loginRetrofitInstance.ssoRegister(request)
-    }
-
-    suspend fun setFcmToken(request: FcmTokenBody): BaseResponse<Unit> {
-        return loginRetrofitInstance.setFcmToken(request)
     }
 }
